Add skip(infinite_t, sentinel) overload to random_access_iterator
This makes it possible to determine the distance between two iterators.  It is a generalisation of skip(count, sentinel) where we don't need a count because we want to hit the sentinel. Existing views and factories have been updated to add this new overload where necessary.

views::concat has been disabled for now because it doesn't support iterating backward when the resultant view is infinite.